Split up builds.js back into builds & commits
